feat!: upgrade Ruby and Bridgetown deployment command
The deploy command is now just `yarn deploy` for new Bridgetown sites. The change to this code reflects that. It also makes it easier to change for the user.

NOTE: you may want to append the env to your deploy command like: `NODE_ENV=production BRIDGETOWN_ENV=production yarn deploy`

This change also bumps the Ruby very to the latest 2.7 release.